Moved MerrMailWorker to Application layer
