--- src/Application/Contracts/IApplicationService.cs
@@ -1,8 +1,0 @@
-namespace Merrsoft.MerrMail.Application.Contracts;
-
-public interface IApplicationService
-{
-    Task<bool> CanStartAsync();
-    Task RunAsync();
-    Task StopAsync();
-}
