+++ src/Domain/Options/DataStorageOptions.cs
@@ -5,8 +5,6 @@ namespace Merrsoft.MerrMail.Domain.Options;
 
 public class DataStorageOptions
 {
-    [Required]
-    public required DataStorageType DatabaseType { get; set; }
-    [Required]
-    public required string DatabaseAccess { get; set; }
+    [Required] public required DataStorageType DatabaseType { get; set; }
+    [Required] public required string DatabaseAccess { get; set; }
 }
